requirements.txt
flask==3.0.0
Werkzeug==3.0.1
prometheus-client==0.18.0
prometheus-flask-exporter==0.23.0
python-dotenv==1.0.0
requests==2.25.1
gunicorn==21.2.0